Fixed issue with loading references
Fixed an issue where loading data from a json file would cause the references would not be loaded as expected. This is because when creating a has h with $refs key in line it will be converted to a symbol but when loading it via json.parse it keeps it as a string.